+++ packages/plugin-divider/web/src/const.ts
@@ -1,5 +1,5 @@
-export const DIVIDER_SIZE_RESOLVER_ID = 'GET_NODE_SIZE';
-export const DIVIDER_STYLE_RESOLVER_ID = 'GET_NODE_STYLE';
+export const DIVIDER_SIZE_RESOLVER_ID = 'GET_DIVIDER_NODE_SIZE';
+export const DIVIDER_STYLE_RESOLVER_ID = 'GET_DIVIDER_NODE_STYLE';
 
 export const DIVIDER_SIZE_BUTTON_DATA_HOOK = 'dividerSizeDropdownButton';
 export const DIVIDER_STYLE_BUTTON_DATA_HOOK = 'dividerStyleDropdownButton';