+++ packages/toolbars-ui/web/src/components/index.ts
@@ -5,3 +5,4 @@ export { ListItemSelect, ListItemSection } from './ListItem';
 export { DropdownPanel } from './DropdownPanel';
 export { AddPluginMenu, PLUGIN_MENU_MODAL_ID } from './AddPluginMenu';
 export { ToolbarButtonSeparator } from './ToolbarButtonSeparator';
+export { ToolbarButton } from './ToolbarButton';